requirements.txt
groq==0.19.0
python-dotenv==1.0.0
requests==2.32.3
imageio[pyav]==2.9.0
imageio[opencv]==2.9.0
pillow==9.2.0
flask==2.0.2
gunicorn==20.1.0
urllib3==2.2.3